+++ CHANGELOG.md
@@ -5,6 +5,29 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Added
+- `SpotifyWebAPIWorker.swift`
+- `SpotifyWebAPIError.swift`
+- `ClientCredsResponsePresets.swift`
+- `SpotifyWebAPIWorkerTests.swift`
+
+### Changed
+- `README.md`
+
+### Deprecated
+TBD
+
+### Removed
+TBD
+
+### Fixed
+TBD
+
+### Security
+TBD
+
 ## [1.0.5] - 2026-02-04
 
 ### Added